NeoNode bugfix: maximum recursion depth (fixes #59)
NeoNode ran into an issue which caused the maximum recursion depth do be
reached (`Could not check message data maximum recursion depth exceeded
while calling a Python object`).

The reason was a ping-pong style recursion buildup between
CheckDataReceived and CheckMessageData. This commit reduced the complexity
of the code and the possibility to run into this recursion buildup.

Currently being tested, final results should be available in a few
hours.