Fix Bug With Edge Case Gas Calc
* If a transaction consumes > 9 GAS and <= 10 GAS, it will not go through because no GAS fee is attached to the transaction. It gets relayed, but never gets processed due to the lack of a fee. The root of this issue comes down to a simple operator issue when determining the GAS fee that is inconsistent with the NEO core reference implementation. The NEO core reference implementation handles this by attached the default "net_fee" of 0.001 when the GAS consumed is <= 10 GAS:

https://github.com/neo-project/neo-gui/blob/master/neo-gui/UI/InvokeContractDialog.cs#L153

The neo-python implementation isn't quite handling this properly due to the fact that it is only checking for GAS consumed being less than zero. In the scenario where the GAS consumed is > 9 and <= 10, the consumed value is -0.xx. The Ceil() converts this to 0, but then the consumed check doesn't properly account for this case. A simple operator change from < to <= solves it.